Move dark mode preferences update out of ThemeProvider
